requirements.txt
numpy==1.26.2
opencv_python==4.8.1.78
Pillow==10.1.0
streamlit==1.28.1
tensorflow==2.14.0
tensorflow_intel==2.14.0
